requirements.txt
asgiref==3.9.1
Django==5.2.4
PyJWT==2.10.1
sqlparse==0.5.3
django-cors-headers
django-redis
